HARVESTER: move harvester crd into model/harvester
