+++ models/harvester/harvesterhci.io.virtualmachineimage.js
