+++ models/harvester/harvesterhci.io.virtualmachinetemplate.js
