+++ models/harvester/harvesterhci.io.virtualmachinetemplateversion.js
