+++ package.json
@@ -73,7 +73,6 @@
     "nuxt": "2.14.6",
     "papaparse": "^5.3.0",
     "portal-vue": "^2.1.5",
-    "randomstring": "^1.2.1",
     "require-extension-hooks": "^0.3.3",
     "require-extension-hooks-babel": "^1.0.0-beta.1",
     "require-extension-hooks-vue": "^3.0.0",
